new code changes

+++ src/main/java/com/coder/lunatic/lunaticEntry/service/LunaticService.java
@@ -1,10 +1,12 @@
 package com.coder.lunatic.lunaticEntry.service;
 
 import com.coder.lunatic.lunaticEntry.entity.LunaticEntry;
+import com.coder.lunatic.lunaticEntry.entity.User;
 import com.coder.lunatic.lunaticEntry.repository.LunaticRepository;
 import org.bson.types.ObjectId;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Component;
+import org.springframework.transaction.annotation.Transactional;
 
 import java.time.LocalDateTime;
 import java.util.List;
@@ -13,10 +15,26 @@ import java.util.Optional;
 @Component
 public class LunaticService {
     @Autowired
-    LunaticRepository repo;
+    private LunaticRepository repo;
 
-    public void saveEntry(LunaticEntry lunaticEntry){
-        lunaticEntry.setDate(LocalDateTime.now());
+    @Autowired
+    private UserService userService;
+
+    @Transactional
+    public void saveEntry(LunaticEntry lunaticEntry, String username){
+        try {
+            User user = userService.findByUserName(username);
+            lunaticEntry.setDate(LocalDateTime.now());
+            LunaticEntry saved = repo.save(lunaticEntry);
+            user.getEntries().add(saved);
+            userService.saveUser(user);
+        } catch (Exception e){
+            System.out.println(e);
+            throw new RuntimeException("An Exception occur "+ e);
+        }
+    }
+
+    public void saveEntry(LunaticEntry lunaticEntry) {
         repo.save(lunaticEntry);
     }
 
@@ -28,8 +46,21 @@ public class LunaticService {
         return repo.findById(id);
     }
 
-    public void deleteById(ObjectId id){
-        repo.deleteById(id);
+    @Transactional
+    public boolean deleteById(ObjectId id, String username){
+        boolean removed = false;
+        try {
+            User user = userService.findByUserName(username);
+            removed = user.getEntries().removeIf(x -> x.getId().equals(id));
+            if (removed) {
+                userService.saveUser(user);
+                repo.deleteById(id);
+            }
+        }catch (Exception e){
+            System.out.println(e);
+            throw new RuntimeException("An exception occur "+ e);
+        }
+        return removed;
     }
 
     public void updateEntry(LunaticEntry LunaticEntry){
